Move product update latency tolerance value object
